% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_report}
\alias{create_report}
\title{Create report}
\usage{
create_report(
  user_ids,
  recent_tweets = 0,
  output_format = html_document(theme = "paper"),
  output_file = "tweep_report.html",
  output_dir = getwd()
)
}
\arguments{
\item{user_ids}{User IDs or screen names of users you want to include
in the report.}

\item{recent_tweets}{Number of recent tweets to include from each user.
Default is 0, so that no tweets are included. I recommend starting
with no recent tweets and then including 3-5 tweets after you've read
the user bios.}

\item{output_format}{Passed to [rmarkdown::render()]. Default value is
`html_document(theme = "paper")`.}

\item{output_file}{Passed to [rmarkdown::render()]. Default is
`"tweep_report.html"`.}

\item{output_dir}{Passed to [rmarkdown::render()]. Default is
`getwd()`, which is the user's current directory.}
}
\description{
Create report
}
\examples{

# summarize the Rohe Lab

rohe_lab <- c("karlrohe",  "krauskae", "alexpghayes", "fchen365",
  "JosephineLukito", "yinizhang2011")

create_report(
  rohe_lab,
  output_file = "rohe_lab_without_recent_tweets.html"
)

create_report(
  rohe_lab,
  recent_tweets = 3,
  output_file = "rohe_lab_with_recent_tweets.html"
)

}
