% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tweet_embed}
\alias{tweet_embed}
\title{Generate HTML to embed a Tweet}
\usage{
tweet_embed(screen_name, status_id, ...)
}
\arguments{
\item{screen_name}{Screen name of the user as a character vector.}

\item{status_id}{Status ID of tweet, as character.}

\item{...}{Parameters to pass to the GET call.}
}
\value{
HTML to embed the tweet as a character vector.
}
\description{
This code shamelessly ripped from
<https://github.com/ropensci/rtweet/pull/305>.
}
\details{
Arguments to pass to the API call can be found at
  <https://developer.twitter.com/en/docs/tweets/post-and-engage/api-reference/get-statuses-oembed>.
}
\examples{

name   <- "kearneymw"
status <- "1087047171306856451"

tweet_embed(screen_name = name, status_id = status)

tweet_embed(
 screen_name = name,
 status_id = status,
 hide_thread = TRUE,
 hide_media = FALSE,
 align = "center"
)

}
\seealso{
[httr::GET()], [httr::content()]
}
